Change CustomSettingsPlayMode
